Commit 4
Added Javascript function and updated the buttons on the home and about pages to represent the GitHub Pages site that is available through the GitHub Pages button.

Javascript has also been added to the site to ensure that mobile devices are supported in terms of the overall navigation and the quality of life for the end user. There has been a hamburger drop down menu that is now supported.